Finish up porting events to their own files
